VoiceAllocator: Bugfix: Saved velocities are floats (#12)
Velocities within the plugin are floats (0 .. 1), so the velsave[]
array where velocity values for notes in the restore_stack
are stored must obviously also contain floats.

+++ plugins/mimid/Engine/VoiceAllocator.h
@@ -120,7 +120,7 @@ private:
 	VoiceList<S> onpri;
 	NoteStack<10> restore_stack;
 	int totalvc;
-	int velsave[128]; // one per note number
+	float velsave[128]; // one per note number
 
 	// Unison mode
 	int uniNote;
